requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
distro==1.7.0
dnspython==2.4.2
docopt==0.6.2
dropbox==11.36.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
netifaces==0.11.0
oauthlib==3.2.0
orjson==3.9.10
packaging==23.2
pillow==10.2.0
ply==3.11
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-extra-types==2.4.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.32.0.post1
stone==3.3.1
termcolor==2.4.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.0.0
yarg==0.1.9
zipp==1.0.0
